feat: 북마크 조회 api JPA로 변환

+++ src/main/java/com/kuit/conet/dto/web/response/team/GetTeamResponseDTO.java
@@ -13,12 +13,12 @@ public class GetTeamResponseDTO {
     private Long teamId;
     private String teamName;
     private String teamImgUrl;
-    private int teamMemberCount;
+    private Long teamMemberCount;
     private Boolean isNew;
     private Boolean bookmark;
 
 
-    public GetTeamResponseDTO(Team team, int teamMemberCount, Boolean bookmark) {
+    public GetTeamResponseDTO(Team team, Long teamMemberCount, Boolean bookmark) {
         this.teamId = team.getId();
         this.teamName = team.getName();
         this.teamImgUrl = team.getImgUrl();
